ETH-3929 convert GET/DELETE body to query params
It's technically legal for GET and DELETE requests to include a request
body, but some frameworks don't bother to support them and will reject
the request. Rude.

Query parameters are supported for all types of requests though, so we
can convert the body to that format when necessary and provide no
request body. Not too tricky.